Validate keys and add tests
